Abstract PreCommitHook
