+++ src/main/scala/io/qbeast/core/model/PreCommitHook.scala
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.qbeast.spark.delta.hook
+package io.qbeast.core.model
 
-import io.qbeast.spark.delta.hook.PreCommitHook.getHookArgName
-import io.qbeast.spark.delta.hook.PreCommitHook.getHookName
-import io.qbeast.spark.delta.hook.PreCommitHook.PreCommitHookOutput
-import org.apache.spark.sql.delta.actions.Action
+import PreCommitHook.getHookArgName
+import PreCommitHook.getHookName
+import PreCommitHook.PreCommitHookOutput
 
 /**
  * A trait representing a pre-commit hook.
@@ -26,7 +25,7 @@ import org.apache.spark.sql.delta.actions.Action
  * Pre-commit hooks are executed before a commit is made to the table. They can be used to perform
  * actions such as validation, logging, or other custom logic.
  */
-trait PreCommitHook {
+trait PreCommitHook[T] {
 
   /**
    * The name of the hook.
@@ -46,7 +45,7 @@ trait PreCommitHook {
    * @return
    *   The output of the hook as a `PreCommitHookOutput`.
    */
-  def run(actions: Seq[Action]): PreCommitHookOutput
+  def run(actions: Seq[T]): PreCommitHookOutput
 
 }
 
@@ -70,33 +69,6 @@ object PreCommitHook {
   def getHookName(hookName: String): String = s"$PRE_COMMIT_HOOKS_PREFIX.$hookName"
 
   def getHookArgName(hookName: String): String = s"${getHookName(hookName)}.$argName"
-
-}
-
-/**
- * A loader for PreCommitHooks
- */
-object QbeastHookLoader {
-
-  /**
-   * Loads a pre-commit hook from a `HookInfo` object.
-   *
-   * This method takes a `HookInfo` object and returns a `PreCommitHook` instance.
-   *
-   * @param hookInfo
-   *   The `HookInfo` object representing the hook to load.
-   * @return
-   *   The loaded `PreCommitHook` instance.
-   */
-  def loadHook(hookInfo: HookInfo): PreCommitHook = hookInfo match {
-    case HookInfo(_, clsFullName, argOpt) =>
-      val cls = Class.forName(clsFullName)
-      val instance =
-        if (argOpt.isDefined)
-          cls.getDeclaredConstructor(argOpt.get.getClass).newInstance(argOpt.get)
-        else cls.getDeclaredConstructor().newInstance()
-      instance.asInstanceOf[PreCommitHook]
-  }
 
 }
