+++ src/test/scala/io/qbeast/spark/delta/hook/QbeastHookLoaderTest.scala
@@ -15,7 +15,8 @@
  */
 package io.qbeast.spark.delta.hook
 
-import io.qbeast.spark.delta.hook.PreCommitHook.PreCommitHookOutput
+import io.qbeast.core.model.HookInfo
+import io.qbeast.core.model.PreCommitHook.PreCommitHookOutput
 import io.qbeast.spark.delta.hook.StatefulTestHook.StatefulTestHookState
 import io.qbeast.spark.internal.QbeastOptions
 import org.apache.spark.sql.delta.actions.Action
@@ -25,7 +26,7 @@ import org.scalatest.matchers.should.Matchers
 
 import java.util.UUID
 
-private class SimpleTestHook extends PreCommitHook {
+private class SimpleTestHook extends DeltaPreCommitHook {
   override val name: String = "SimpleTestHook"
 
   var args: Seq[Action] = Seq.empty
@@ -37,7 +38,7 @@ private class SimpleTestHook extends PreCommitHook {
 
 }
 
-private class StatefulTestHook(val stateId: String) extends PreCommitHook {
+private class StatefulTestHook(val stateId: String) extends DeltaPreCommitHook {
 
   val state: StatefulTestHookState = StatefulTestHook.stateMap(stateId)
 
@@ -73,9 +74,9 @@ class QbeastHookLoaderTest extends AnyFlatSpec with Matchers {
     when(qbeastOptions.hookInfo).thenReturn(
       HookInfo("", classOf[SimpleTestHook].getCanonicalName, None) :: Nil)
 
-    val hookOpts = qbeastOptions.hookInfo.map(QbeastHookLoader.loadHook)
+    val hookOpts = qbeastOptions.hookInfo.map(DeltaHookLoader.loadHook)
     hookOpts.size shouldBe 1
-    hookOpts.head shouldBe a[PreCommitHook]
+    hookOpts.head shouldBe a[DeltaPreCommitHook]
 
     val mockActions = mock(classOf[List[Action]])
     hookOpts.head.run(mockActions)
@@ -90,9 +91,9 @@ class QbeastHookLoaderTest extends AnyFlatSpec with Matchers {
       when(qbeastOptions.hookInfo).thenReturn(
         HookInfo("", classOf[StatefulTestHook].getCanonicalName, Some(argument)) :: Nil)
 
-      val hooks = qbeastOptions.hookInfo.map(QbeastHookLoader.loadHook)
+      val hooks = qbeastOptions.hookInfo.map(DeltaHookLoader.loadHook)
       hooks.size shouldBe 1
-      hooks.head shouldBe a[PreCommitHook]
+      hooks.head shouldBe a[DeltaPreCommitHook]
 
       val mockActions = mock(classOf[List[Action]])
       hooks.head.run(mockActions)
@@ -106,7 +107,7 @@ class QbeastHookLoaderTest extends AnyFlatSpec with Matchers {
     val qbeastOptions = mock(classOf[QbeastOptions])
     when(qbeastOptions.hookInfo).thenReturn(Nil)
 
-    val hookOpts = qbeastOptions.hookInfo.map(QbeastHookLoader.loadHook)
+    val hookOpts = qbeastOptions.hookInfo.map(DeltaHookLoader.loadHook)
     hookOpts.size shouldBe 0
   }
