+++ src/test/scala/io/qbeast/spark/internal/sources/QbeastOptionsTest.scala
@@ -15,8 +15,8 @@
  */
 package io.qbeast.spark.internal.sources
 
-import io.qbeast.spark.delta.hook.HookInfo
-import io.qbeast.spark.delta.hook.PreCommitHook.PRE_COMMIT_HOOKS_PREFIX
+import io.qbeast.core.model.HookInfo
+import io.qbeast.core.model.PreCommitHook.PRE_COMMIT_HOOKS_PREFIX
 import io.qbeast.spark.internal.QbeastOptions
 import io.qbeast.spark.QbeastIntegrationTestSpec
 import org.apache.spark.qbeast.config
